Fix for handling bad populated refs
Restores old behavior.  E.g. if suffix == "_obj", and we populate a
field called "user", which points to a nonexistent object, "user" is
replaced with null, as it was previously.

This is good when interacting with view layers like ExtJS which will
try to populate that ID, and fail because it doesn't exist.